requirements.txt
aiogram==3.16.0
aiohttp==3.9.5
asyncio==3.4.3
asyncpg==0.30.0
DateTime==5.5
loguru==0.7.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
backoff==2.2.1